+++ src/components/ListResPost.tsx
@@ -7,7 +7,7 @@ import dot from '@/data/images/Ellipse 14.svg';
 import { useRouter } from 'next/navigation';
 import { topicMapping } from '@/utils/topics';
 import { Post } from '@/app/home/_types/homeTypes';
-import { useGPTTranslation } from '@/app/post/_hooks/TranslatedText';
+import { useGPTTranslation } from '@/app/post/_hooks/useGPTTranslation';
 import RenderonlyTextHTML from '@/hook/home/RenderonlyTextHTML';
 import TimeAgo from '@/app/search/[id]/_components/TimeAgo';
